# internal/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for reading and writing Branca-encoded data on Mifare Classic NFC cards using D8/T8 hardware.",
        "title": "LTO NFC Card Reader API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/device/beep": {
            "post": {
                "description": "Triggers the device buzzer for the specified duration in milliseconds",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Trigger buzzer",
                "parameters": [
                    {
                        "description": "Beep request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DeviceBeepRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/device/eeprom/read": {
            "post": {
                "description": "Reads bytes from the device EEPROM at the specified offset (offset: 0–383, length: 1–384)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Read EEPROM",
                "parameters": [
                    {
                        "description": "EEPROM read request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EEPROMReadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.EEPROMReadResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/device/eeprom/write": {
            "post": {
                "description": "Writes hex-encoded bytes to the device EEPROM at the specified offset (offset: 0–383)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Write EEPROM",
                "parameters": [
                    {
                        "description": "EEPROM write request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EEPROMWriteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/device/reset": {
            "post": {
                "description": "Resets the RF field for the specified duration in milliseconds",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "RF field reset",
                "parameters": [
                    {
                        "description": "Reset request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DeviceResetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/device/value/decrement": {
            "post": {
                "description": "Subtracts the given value from a Mifare value block (block: 1–63)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Decrement value block",
                "parameters": [
                    {
                        "description": "Decrement request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DecrementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/device/value/increment": {
            "post": {
                "description": "Adds the given value to a Mifare value block (block: 1–63)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Increment value block",
                "parameters": [
                    {
                        "description": "Increment request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.IncrementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/device/value/init": {
            "post": {
                "description": "Formats a card block as a Mifare value block with the specified initial value (block: 1–63)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Initialize value block",
                "parameters": [
                    {
                        "description": "Init value request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InitValRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/device/value/read": {
            "post": {
                "description": "Reads the current integer value from a Mifare value block (block: 0–63)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Read value block",
                "parameters": [
                    {
                        "description": "Read value request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReadValRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ReadValResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.DecrementRequest": {
            "type": "object",
            "properties": {
                "block": {
                    "description": "Block address",
                    "type": "integer"
                },
                "value": {
                    "description": "Value to subtract",
                    "type": "integer"
                }
            }
        },
...
        "models.DeviceBeepRequest": {
            "type": "object",
            "properties": {
                "ms": {
                    "description": "Duration in milliseconds",
                    "type": "integer"
                }
            }
        },
        "models.DeviceResetRequest": {
            "type": "object",
            "properties": {
                "ms": {
                    "description": "Reset duration in milliseconds",
                    "type": "integer"
                }
            }
        },
        "models.EEPROMReadRequest": {
            "type": "object",
            "properties": {
                "length": {
                    "description": "1-384",
                    "type": "integer"
                },
                "offset": {
                    "description": "0-383",
                    "type": "integer"
                }
            }
        },
        "models.EEPROMReadResponse": {
            "type": "object",
            "properties": {
                "bytes": {
                    "description": "Raw byte array",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "data": {
                    "description": "Hex string",
                    "type": "string"
                }
            }
        },
        "models.EEPROMWriteRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Hex string",
                    "type": "string"
                },
                "offset": {
                    "description": "0-383",
                    "type": "integer"
                }
            }
        },
        "models.IncrementRequest": {
            "type": "object",
            "properties": {
                "block": {
                    "description": "Block address",
                    "type": "integer"
                },
                "value": {
                    "description": "Value to add",
                    "type": "integer"
                }
            }
        },
        "models.InitValRequest": {
            "type": "object",
            "properties": {
                "block": {
                    "description": "Block address",
                    "type": "integer"
                },
                "value": {
                    "description": "Initial value",
                    "type": "integer"
                }
            }
        },
        "models.ReadValRequest": {
            "type": "object",
            "properties": {
                "block": {
                    "description": "Block address",
                    "type": "integer"
                }
            }
        },
        "models.ReadValResponse": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "integer"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
...
}